+++ src/services/http/middleware/I18n.ts
@@ -1,62 +1,29 @@
 import type { NextFunction, Response } from 'express';
-import i18next, { type i18n, type TFunction } from 'i18next';
-import BackendFS from 'i18next-fs-backend';
 import type i18nConfig from '../../../config/i18n.ts';
 import type { IApp } from '../../../server.ts';
+import type { TI18n } from '../../i18n/I18n.ts';
 import type { FrameworkRequest } from '../HttpServer.ts';
 import AbstractMiddleware from './AbstractMiddleware.ts';
 import type { GetUserByTokenAppInfo } from './GetUserByToken.ts';
 
-export type TI18n = { t: TFunction; language: string };
+export type { TI18n };
 
 export interface I18nMiddlewareAppInfo {
   appInfo: {
-    i18n: {
-      t: TFunction;
-      language: string;
-    };
+    i18n: TI18n;
   };
 }
 
 class I18n extends AbstractMiddleware {
-  cache: { [key: string]: i18n } = {};
-
   enabled = true;
 
   lookupQuerystring = '';
 
-  supportedLngs: Array<string> = [];
-
-  fallbackLng = 'en';
-
-  i18n: { t: TFunction; language: string } = {
-    t: ((text) => text) as TFunction,
-    language: 'en',
-  };
-
   constructor(app: IApp, params?: Record<string, unknown>) {
     super(app, params);
     const I18NConfig = this.app.getConfig('i18n') as typeof i18nConfig;
-
-    if (I18NConfig.enabled) {
-      this.logger?.info('Enabling i18n support');
-      this.i18n = i18next;
-      i18next.use(BackendFS).init({
-        backend: {
-          loadPath: `${this.app.foldersConfig.locales}/{{lng}}/{{ns}}.json`,
-          addPath: `${this.app.foldersConfig.locales}/{{lng}}/{{ns}}.missing.json`,
-        },
-        fallbackLng: I18NConfig.fallbackLng,
-        preload: I18NConfig.preload,
-        saveMissing: I18NConfig.saveMissing,
-        debug: I18NConfig.debug,
-      });
-    }
-
     this.enabled = I18NConfig.enabled;
     this.lookupQuerystring = I18NConfig.lookupQuerystring;
-    this.supportedLngs = I18NConfig.supportedLngs;
-    this.fallbackLng = I18NConfig.fallbackLng;
   }
 
   static get description() {
@@ -64,31 +31,14 @@ class I18n extends AbstractMiddleware {
   }
 
   async middleware(req: FrameworkRequest, _res: Response, next: NextFunction) {
-    let i18n: i18n | TI18n | null = null;
+    let lang = '';
 
     if (this.enabled) {
-      let lang = this.detectLang(req);
-      if (!lang || this.supportedLngs.indexOf(lang) === -1) {
-        this.logger?.verbose(
-          `Language "${lang}" is not supported or not detected. Using fallback on ${this.fallbackLng}`,
-        );
-        lang = this.fallbackLng;
-      }
-
-      if (!this.cache[lang]) {
-        this.cache[lang] = i18next.cloneInstance({
-          initAsync: false,
-          lng: lang,
-        });
-      }
-      i18n = this.cache[lang];
+      lang = await this.detectLang(req);
     }
+    const i18nService = await this.app.getI18nService();
 
-    if (!i18n) {
-      i18n = this.i18n;
-    }
-
-    req.appInfo.i18n = i18n;
+    req.appInfo.i18n = await i18nService.getI18n(lang);
     //@ts-expect-error we known thats a new one
     req.i18n = new Proxy(req.appInfo.i18n, {
       get: (target, prop) => {
@@ -116,16 +66,22 @@ class I18n extends AbstractMiddleware {
 
   detectorOrder = ['XLang', 'query', 'user'];
 
-  detectLang(req: FrameworkRequest, isUseShortCode = true): string {
+  async detectLang(
+    req: FrameworkRequest,
+    isUseShortCode = true,
+  ): Promise<string> {
     let lang = '';
     for (const detectorName of this.detectorOrder) {
       const lng = this.detectors[detectorName](req);
       if (!lng) {
         continue;
       }
-      if (i18next.services.languageUtils.isSupportedCode(lng)) {
+      const i18nService = await this.app.getI18nService();
+      const i18nInstance = await i18nService.getI18nInstance();
+      if (i18nInstance.services.languageUtils.isSupportedCode(lng)) {
         if (isUseShortCode) {
-          lang = i18next.services.languageUtils.getLanguagePartFromCode(lng);
+          lang =
+            i18nInstance.services.languageUtils.getLanguagePartFromCode(lng);
         } else {
           lang = lng;
         }